+++ internal/io/sfarcio/sfarcio.go
@@ -2,6 +2,7 @@ package sfarcio
 
 import (
 	"database/sql"
+	"log/slog"
 
 	"github.com/sfborg/from-dwca/internal/ent/sfarc"
 	"github.com/sfborg/from-dwca/pkg/config"
@@ -25,14 +26,16 @@ func (s *sfarcio) Exists() bool {
 		return false
 	}
 
-	q := "SELECT dwc_taxon_id FROM core LIMIT 1"
+	q := "SELECT dwc_taxon_id FROM core LIMIT 5"
 
 	var id string
 	err := s.db.QueryRow(q).Scan(&id)
 	if err != nil {
+		slog.Error("Cannot get data from core", "error", err)
 		return false
 	}
 	if id == "" {
+		slog.Error("No dwc_taxon_id in core")
 		return false
 	}