+++ pkg/fdwca.go
@@ -4,18 +4,18 @@ import (
 	dwca "github.com/gnames/dwca/pkg"
 	dwcacfg "github.com/gnames/dwca/pkg/config"
 	"github.com/gnames/gnparser"
-	"github.com/sfborg/from-dwca/internal/ent/sfarc"
+	"github.com/sfborg/from-coldp/pkg/ent/sfgarc"
 	"github.com/sfborg/from-dwca/pkg/config"
 )
 
 type fdwca struct {
 	cfg     config.Config
-	s       sfarc.Archive
+	s       sfgarc.Archive
 	d       dwca.Archive
 	gnpPool chan gnparser.GNparser
 }
 
-func New(cfg config.Config, sfarc sfarc.Archive) FromDwCA {
+func New(cfg config.Config, sfgarc sfgarc.Archive) FromDwCA {
 	res := &fdwca{cfg: cfg}
 
 	poolSize := cfg.JobsNum
@@ -25,7 +25,7 @@ func New(cfg config.Config, sfarc sfarc.Archive) FromDwCA {
 		gnpPool <- gnparser.New(cfgGNP)
 	}
 	res.gnpPool = gnpPool
-	res.s = sfarc
+	res.s = sfgarc
 
 	return res
 }
@@ -33,7 +33,7 @@ func New(cfg config.Config, sfarc sfarc.Archive) FromDwCA {
 func (fd *fdwca) GetDwCA(fileDwCA string) (dwca.Archive, error) {
 	opts := []dwcacfg.Option{
 		dwcacfg.OptJobsNum(fd.cfg.JobsNum),
-		dwcacfg.OptWrongFieldsNum(fd.cfg.WrongFieldsNum),
+		dwcacfg.OptWrongFieldsNum(fd.cfg.BadRow),
 	}
 
 	dwcaCfg := dwcacfg.New(opts...)
@@ -66,7 +66,7 @@ func (fd *fdwca) GetDwCA(fileDwCA string) (dwca.Archive, error) {
 
 func (fd *fdwca) ImportDwCA(arc dwca.Archive) error {
 	fd.d = arc
-	num, err := fd.importCore()
+	err := fd.importCore()
 	if err != nil {
 		return err
 	}
@@ -76,7 +76,7 @@ func (fd *fdwca) ImportDwCA(arc dwca.Archive) error {
 		return err
 	}
 
-	err = fd.importEML(arc.EML(), num)
+	err = fd.importEML(arc.EML())
 	if err != nil {
 		return err
 	}
@@ -91,8 +91,4 @@ func (f *fdwca) ExportSFGA(outputPath string) error {
 	}
 
 	return nil
-}
-
-func (fd *fdwca) checkSFGA() bool {
-	return fd.s.Exists()
 }