+++ pkg/import_eml.go
@@ -3,25 +3,23 @@ package fdwca
 import (
 	"strings"
 
+	"github.com/gnames/coldp/ent/coldp"
 	"github.com/gnames/dwca/pkg/ent/eml"
-	"github.com/sfborg/from-dwca/internal/ent/schema"
 )
 
-func (fd *fdwca) importEML(data *eml.EML, recNum int) error {
+func (fd *fdwca) importEML(data *eml.EML) error {
 	var doi string
 	if strings.Contains(data.Dataset.ID, "doi.") {
 		doi = data.Dataset.ID
 	}
 
-	ds := schema.DataSource{
-		ID:          data.Dataset.ID,
+	meta := coldp.Meta{
 		Title:       data.Dataset.Title,
 		DOI:         doi,
 		Description: data.Dataset.Abstract.Para,
-		RecordCount: recNum,
 	}
 
-	err := fd.s.InsertDataSource(&ds)
+	err := fd.s.InsertMeta(&meta)
 	if err != nil {
 		return err
 	}